backend/requirements.txt
gunicorn
uvicorn
fastapi
strawberry-graphql
python-dotenv
pydantic
sqlmodel
greenlet
asyncpg
SQLAlchemy
PyJWT
colorlog
